backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
supabase==2.10.0
PyMuPDF==1.24.14
pydantic-settings==2.6.1
httpx==0.27.2
